Module 03/requirements.txt
azure-functions==1.18.0
azure-storage-blob==12.19.0
azure-cosmos==4.5.1
azure-keyvault-secrets==4.7.0
azure-identity==1.15.0
Pillow==10.1.0
python-magic==0.4.27
requests==2.31.0
